refactor: remove global context aggregator
