+++ Frontend-nextjs/app/(protected)/home/cards/ProssimoPagamentoCard.tsx
@@ -6,7 +6,7 @@ import { CalendarCheck } from "lucide-react";
 
 import DashboardCard from "./DashboardCard";
 import LoadingSpinnerCard from "./loading/LoadingSpinnerCard";
-import { useRicorrenze } from "@/context/contexts/RicorrenzeContext";
+import { useRicorrenze } from "@/context/RicorrenzeContext";
 import { useRenderTimer } from "@/app/(protected)/home/utils/useRenderTimer"; // Debug per vedere quanto tempo ci mette a rtenderizzare
 
 export default function ProssimoPagamentoCard() {
