+++ Frontend-nextjs/app/(protected)/profilo/ProfilePage.tsx
@@ -11,8 +11,8 @@ import ProfileRow from "./components/ProfileRow";
 import ThemeSelectorRow from "./components/ThemeSelectorRow";
 import AvatarPickerModal from "./components/AvatarPickerModal";
 import { DEFAULT_USER, UserType } from "@/types/models/user";
-import { useUser } from "@/context/contexts/UserContext";
-import { useThemeContext } from "@/context/contexts/ThemeContext";
+import { useUser } from "@/context/UserContext";
+import { useThemeContext } from "@/context/ThemeContext";
 import PendingEmailNotice from "@/app/components/PendingEmailNotice";
 import { UserRound } from "lucide-react";
 import LegalLinks from "@/app/components/legal/LegalLinks";
