+++ Frontend-nextjs/app/(protected)/transazioni/layout.tsx
@@ -1,0 +1,5 @@
+import { SelectionProvider } from "@/context/SelectionContext";
+
+export default function TransazioniLayout({ children }: { children: React.ReactNode }) {
+    return <SelectionProvider>{children}</SelectionProvider>;
+}
