+++ Frontend-nextjs/context/CategoriesContext.tsx
