+++ Frontend-nextjs/src/context/CategoriesContext.ts
@@ -2,4 +2,4 @@
 // Sezione: Re-export CategoriesContext
 // Dettagli: espone hook useCategories dal contesto principale
 // ─────────────────────────────────────────────────────────────────────────────
-export { useCategories } from '@/context/contexts/CategoriesContext';
+export { useCategories } from '@/context/CategoriesContext';
